pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "world-retail-congress-scraper"
version = "1.0.0"
description = "World Retail Congress Speakers Scraper with Streamlit UI"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "playwright>=1.40.0",
    "requests>=2.31.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.30.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "plotly>=5.18.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/world-retail-congress-scraper"
"Bug Tracker" = "https://github.com/yourusername/world-retail-congress-scraper/issues"

[project.scripts]
wrc-scraper = "app:main"

[tool.setuptools]
packages = ["models", "views", "controllers", "utils"]

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest]
testpaths = ["tests"]